+++ src/buzzer.h
@@ -1,0 +1,7 @@
+#ifndef BUZZER_H
+#define BUZZER_H
+
+void playStartupChime();
+void playAlertChime();
+
+#endif // BUZZER_H
