+++ src/display.h
@@ -1,0 +1,8 @@
+#ifndef DISPLAY_H
+#define DISPLAY_H
+
+void displayClock();
+void haltAndCatchFire();
+void displayScrollingInfo();
+
+#endif // DISPLAY_H
