+++ src/main.cpp
@@ -12,14 +12,11 @@
 #include <WiFiManager.h> // https://github.com/tzapu/WiFiManager
 #include <PubSubClient.h> // https://github.com/knolleary/pubsubclient
 #include <ArduinoJson.h> // For JSON serialization
-
-// --- Existing Defines ---
-#define INDOOR 0x12
-#define OUTDOOR 0xE
-#define LIGHTNING_INT 0x08
-#define DISTURBER_INT 0x04
-#define NOISE_INT 0x01
-#define BUZZER_PIN 13
+#include "globals.h"
+#include "buzzer.h"
+#include "display.h"
+#include "mqtt_utils.h"
+#include "webserver_utils.h"
 
 // --- LCD and Sensor Objects ---
 LiquidCrystal_I2C lcd(0x27, 16, 2); // Address, 16 chars, 2 lines
@@ -94,311 +91,6 @@ WiFiManagerParameter custom_mqtt_topic_status("top_s", "status topic", mqtt_topi
 
 unsigned long lastMQTTPublishMillis = 0;
 const unsigned long mqttPublishInterval = 5UL * 60UL * 1000UL; // Publish every 5 minutes
-
-// --- Forward Declarations (best practice) ---
-void playStartupChime();
-void playAlertChime();
-void displayClock();
-void haltAndCatchFire();
-void displayScrollingInfo();
-void saveCustomParamsCallback();
-bool testMqttConnection(const char* server_addr, int port, const char* user, const char* pass);
-void reconnectMqtt();
-void mqttCallback(char* topic, byte* payload, unsigned int length);
-void handleRoot();
-void handleJsonStatus();
-void handleConfig(); // NEW: Forward declaration for the config handler
-
-
-// --- BUZZER FUNCTIONS ---
-void playStartupChime()
-{
-  tone(BUZZER_PIN, 523, 150);
-  delay(200);
-  tone(BUZZER_PIN, 659, 150);
-  delay(200);
-  noTone(BUZZER_PIN);
-}
-
-void playAlertChime()
-{
-  tone(BUZZER_PIN, 523, 200);
-  delay(250);
-  tone(BUZZER_PIN, 440, 200);
-  delay(250);
-  tone(BUZZER_PIN, 659, 250);
-  delay(300);
-  noTone(BUZZER_PIN);
-}
-
-
-// --- DISPLAY FUNCTIONS ---
-void displayClock()
-{
-  DateTime now = rtc.now();
-  char timeStr[6];
-  sprintf(timeStr, "%02d:%02d", now.hour(), now.minute()); // Still using 24-hour for compactness
-
-  int startCol = 15 - strlen(timeStr); // right-align
-  lcd.setCursor(1, 0); // Start clearing from a safe spot, not column 0 where WiFi icon is
-  lcd.print("               "); // Clear the rest of the line
-  lcd.setCursor(startCol, 0);
-  lcd.print(timeStr);
-}
-
-
-void haltAndCatchFire()
-{
-  while (1)
-  {
-    playAlertChime();
-    delay(1500); // Wait for 1.5 seconds
-  };
-}
-
-void displayScrollingInfo()
-{
-  lcd.setCursor(0, 1);
-  switch (displayState)
-  {
-  case 0:
-    lcd.write((uint8_t)2); // thermometer
-    lcd.print(" Temp: ");
-    lcd.print(currentTempF, 1); // Display Fahrenheit
-    lcd.print("F   ");
-    break;
-  case 1:
-    lcd.write((uint8_t)4); // pressure (using the new pressure icon)
-    lcd.print(" Pres:");
-    lcd.print(currentPressureInHg, 2); // Display inHg with 2 decimal places
-    lcd.print("inHg");
-    break;
-  case 2:
-    lcd.write((uint8_t)1); // lightning bolt
-    lcd.print(" Last: ");
-    if (lastDistanceMiles >= 0)
-    {
-      lcd.print(lastDistanceMiles);
-      lcd.print("mi   "); // Change unit to miles
-    }
-    else
-    {
-      lcd.print("None     ");
-    }
-    break;
-  case 3:
-    lcd.write((uint8_t)1); // lightning bolt
-    lcd.print(" Energy: ");
-    if (lastEnergy >= 0)
-    {
-      lcd.print(lastEnergy);
-      lcd.print("      ");
-    }
-    else
-    {
-      lcd.print("None     ");
-    }
-    break;
-  default:
-    displayState = 0;
-    return;
-  }
-  displayState = (displayState + 1) % 4;
-}
-
-// --- WiFiManager Custom Parameter Save Callback ---
-void saveCustomParamsCallback() {
-  Serial.println("Should save config");
-  strcpy(mqtt_server, custom_mqtt_server.getValue());
-  strcpy(mqtt_port, custom_mqtt_port.getValue());
-  strcpy(mqtt_username, custom_mqtt_username.getValue());
-  strcpy(mqtt_password, custom_mqtt_password.getValue());
-  strcpy(mqtt_topic_temp, custom_mqtt_topic_temp.getValue());
-  strcpy(mqtt_topic_pressure, custom_mqtt_topic_pressure.getValue());
-  strcpy(mqtt_topic_lightning_dist, custom_mqtt_topic_lightning_dist.getValue());
-  strcpy(mqtt_topic_lightning_energy, custom_mqtt_topic_lightning_energy.getValue());
-  strcpy(mqtt_topic_status, custom_mqtt_topic_status.getValue());
-
-  Serial.print("MQTT Server: "); Serial.println(mqtt_server);
-  Serial.print("MQTT Port: "); Serial.println(mqtt_port);
-  // Do not print password for security reasons
-}
-
-// --- MQTT Test Function for WiFiManager ---
-bool testMqttConnection(const char* server_addr, int port, const char* user, const char* pass) { // Renamed 'server' to 'server_addr' to avoid conflict with WebServer object
-  Serial.print("Attempting MQTT test connect to "); Serial.print(server_addr); Serial.print(":"); Serial.println(port);
-  lcd.setCursor(0,0);
-  lcd.print("Testing MQTT...");
-  lcd.setCursor(0,1);
-  lcd.print("               "); // Clear line 2
-
-  mqttClient.setServer(server_addr, port);
-  // Generate a unique client ID
-  String clientId = "ESP32Weather-";
-  clientId += String(micros(), HEX); // Using micros() for a somewhat unique ID
-
-  bool connected;
-  if (strlen(user) > 0) {
-    connected = mqttClient.connect(clientId.c_str(), user, pass);
-  } else {
-    connected = mqttClient.connect(clientId.c_str());
-  }
-
-  if (connected) {
-    Serial.println("MQTT Test CONNECTED");
-    lcd.setCursor(0,1);
-    lcd.print("MQTT OK!      ");
-    mqttClient.disconnect(); // Disconnect immediately after successful test
-    delay(1000); // Give user time to see success
-    return true;
-  } else {
-    Serial.print("MQTT Test FAILED, rc=");
-    Serial.println(mqttClient.state());
-    lcd.setCursor(0,1);
-    lcd.print("MQTT Failed! rc:");
-    lcd.print(mqttClient.state());
-    delay(2000); // Give user time to see failure
-    return false;
-  }
-}
-
-// --- MQTT Callback Function (for subscribed topics, if any) ---
-void mqttCallback(char* topic, byte* payload, unsigned int length) {
-  Serial.print("Message arrived [");
-  Serial.print(topic);
-  Serial.print("] ");
-  for (int i = 0; i < length; i++) {
-    Serial.print((char)payload[i]);
-  }
-  Serial.println();
-  // Add logic here to act on received MQTT messages if needed
-}
-
-// --- MQTT Reconnection Logic ---
-void reconnectMqtt() {
-  // Loop until we're reconnected
-  while (!mqttClient.connected()) {
-    Serial.print("Attempting MQTT connection...");
-    lcd.setCursor(0,1);
-    lcd.print("MQTT Reconn...");
-    // Generate a unique client ID based on MAC address for persistence
-    String clientId = "ESP32Weather-";
-    clientId += WiFi.macAddress().substring(12); // Last 2 bytes of MAC for a shorter ID
-
-    // Attempt to connect
-    bool connected;
-    if (strlen(mqtt_username) > 0) {
-      connected = mqttClient.connect(clientId.c_str(), mqtt_username, mqtt_password);
-    } else {
-      connected = mqttClient.connect(clientId.c_str());
-    }
-
-    if (connected) {
-      Serial.println("MQTT Connected!");
-      lcd.setCursor(0,1);
-      lcd.print("MQTT Connected  ");
-      // Subscribe to any topics if needed, e.g., for commands
-      // mqttClient.subscribe("weather/commands");
-      mqttClient.publish(mqtt_topic_status, "Weather Station Online"); // Publish online status
-      lastMQTTPublishMillis = millis(); // Reset publish timer on connect
-    } else {
-      Serial.print("MQTT failed, rc=");
-      Serial.print(mqttClient.state());
-      Serial.println(" try again in 5 seconds");
-      lcd.setCursor(0,1);
-      lcd.print("MQTT Failed, ");
-      lcd.print(mqttClient.state());
-      delay(5000); // Wait 5 seconds before retrying
-    }
-  }
-}
-
-// --- Web Server Handlers ---
-void handleRoot() {
-  String html = R"rawliteral(
-    <!DOCTYPE html>
-    <html>
-    <head>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>ESP32 Weather Station</title>
-    <style>
-      body { font-family: Arial, Helvetica, sans-serif; text-align: center; }
-      h1 { color: #333; }
-      .container { max-width: 500px; margin: auto; padding: 20px; border: 1px solid #ddd; border-radius: 8px; box-shadow: 0 2px 4px rgba(0,0,0,0.1); }
-      p { line-height: 1.6; }
-      a { color: #007bff; text-decoration: none; }
-      a:hover { text-decoration: underline; }
-      code { background-color: #f4f4f4; padding: 2px 4px; border-radius: 4px; }
-    </style>
-    </head>
-    <body>
-    <div class="container">
-      <h1>ESP32 Weather Station</h1>
-      <p>Your weather station is up and running!</p>
-      <p>View sensor data in JSON format: <a href="/json">/json</a></p>
-      <p><b><a href="/config">Configure WiFi/MQTT Settings</a></b></p>
-      <p>Current IP: <code>)rawliteral";
-  html += WiFi.localIP().toString();
-  html += R"rawliteral(</code></p>
-    </div>
-    </body>
-    </html>
-  )rawliteral";
-  server.send(200, "text/html", html);
-}
-
-void handleJsonStatus() {
-  StaticJsonDocument<512> doc;
-
-  // Add sensor readings
-  doc["temperature_F"] = currentTempF;
-  doc["pressure_inHg"] = currentPressureInHg;
-
-  // Add lightning strike info
-  if (lastDistanceMiles >= 0) {
-    doc["last_lightning_distance_miles"] = lastDistanceMiles;
-    doc["last_lightning_energy"] = lastEnergy;
-  } else {
-    doc["last_lightning_distance_miles"] = "N/A";
-    doc["last_lightning_energy"] = "N/A";
-  }
-
-  // Add clock time
-  DateTime now = rtc.now();
-  char timeStr[20];
-  sprintf(timeStr, "%02d:%02d:%02d", now.hour(), now.minute(), now.second());
-  doc["current_time"] = timeStr;
-
-  char dateStr[20];
-  sprintf(dateStr, "%04d-%02d-%02d", now.year(), now.month(), now.day());
-  doc["current_date"] = dateStr;
-
-  // Add network status
-  doc["wifi_status"] = WiFi.status() == WL_CONNECTED ? "connected" : "disconnected";
-  doc["ip_address"] = WiFi.localIP().toString();
-  doc["mqtt_status"] = mqttClient.connected() ? "connected" : "disconnected";
-
-  String jsonString;
-  serializeJson(doc, jsonString);
-
-  server.send(200, "application/json", jsonString);
-}
-
-// NEW: Handler to launch the WiFiManager configuration portal
-void handleConfig() {
-  server.send(200, "text/html", "Launching configuration portal... You will be redirected to the AP.");
-  Serial.println("Manually launching WiFiManager configuration portal...");
-  lcd.clear();
-  lcd.setCursor(0,0);
-  lcd.print("Config Mode...");
-  lcd.setCursor(0,1);
-  lcd.print("Connect to AP:");
-  delay(100); // Small delay to allow response to send to client before AP starts
-  wm.startConfigPortal("WeatherStationSetup"); // Re-launch the config portal
-  // After configuration, WiFiManager will restart the ESP32.
-  // This function won't return until the portal is exited and reboot happens.
-}
-
 
 // --- SETUP FUNCTION ---
 void setup()
