+++ src/webserver_utils.h
@@ -1,0 +1,8 @@
+#ifndef WEBSERVER_UTILS_H
+#define WEBSERVER_UTILS_H
+
+void handleRoot();
+void handleJsonStatus();
+void handleConfig();
+
+#endif // WEBSERVER_UTILS_H
